refactor app mapping
