+++ webapps/app-st/src/App.py
@@ -1,8 +1,8 @@
 import streamlit as st
 
 from components.auth.authenticated import authenticated
-from states.app import init_app_state_mapping
 from states.auth import get_user
+from states.mappings import init_app_state_mapping
 
 
 st.set_page_config(
