+++ webapps/app-st/src/states/mappings.py
@@ -2,7 +2,7 @@
 Mapeos de llaves de estados de la aplicación en `st.session_state`
 """
 
-from typing import Any
+from typing import Callable, Any
 
 import streamlit as st
 from pydantic import BaseModel
@@ -36,11 +36,28 @@ class AppMap(StateMap):
         }
 
 
-def map_session_state(sm: StateMap):
+
+def _map_session_state(sm: StateMap):
     defaults: dict[str, Any] = sm.get_default_state_values()
     for field, mapping in sm:
-        print(f"field={field}, mapping={mapping}")
         if isinstance(mapping, StateMap):
-            map_session_state(mapping)
+            _map_session_state(mapping)
         elif field in defaults:
             st.session_state[mapping] = defaults[field]
+
+
+@st.cache_data
+def get_app_state_mapping() -> AppMap:
+    return AppMap()
+
+
+def init_app_state_mapping(fn: Callable[..., Any]) -> Callable[..., Any]:
+    def wrapper(*args: Any, **kwargs: Any) -> Any:
+        if 'app' not in st.session_state:
+            _map_session_state(get_app_state_mapping())
+            st.session_state['app'] = True
+
+        return fn(*args, **kwargs)
+
+    return wrapper
+
